pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "movies-api"
version = "0.0.1"
description = "API for movies"
readme = "README.md"
authors = [ "Henrique Pasquini Santos <riquepasq@gmail.com>",]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "3.11.10"
elasticsearch = "<8"
fastapi = "^0.115.5"
google-cloud-redis = "^2.16.1"
python-decouple = "^3.8"
python-json-logger = "^2.0.7"
redis = "^3.5.3"
redis-py-cluster = "^2.1.3"
regex = "^2024.9.11"
requests = "^2.32.3"
uvicorn = "^0.32.1"

[tool.poetry.group.doc.dependencies]
toml = "^0.10.2"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"
black = "^24.8.0"
flake8 = "^7.1.1"
pylint = "^3.3.1"
pytest = "^8.2.1"
pytest-cov = "^4.1"
pytest-xdist = "^3.6.1"
pytest-asyncio = "^1.1.0"
